moved handleForm into utils, updated animations
