+++ src/components/Banner/Banner.module.scss
@@ -7,6 +7,7 @@
   width: 100%;
   min-height: 400px;
   color: $white;
+  background-repeat: no-repeat;
 
   @include breakpoint(lg) {
     min-height: 500px;
